requirement.txt
alembic==1.4.1
aniso8601==8.0.0
autopep8==1.5
blurhash-python==1.0.1
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
